Move screenings model to lib/mongodb/models
